# review-service/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/delete/{reviewID}:
        delete:
            tags:
                - Review
            description: c端删除评价
            operationId: Review_DeleteReview
            parameters:
                - name: reviewID
                  in: path
                  required: true
                  schema:
                    type: string
                - name: userID
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.DeleteReviewReply'
    /v1/review:
        post:
            tags:
                - Review
            description: c端创建评价
            operationId: Review_CreateReview
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.review.v1.CreateReviewRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.CreateReviewReply'
    /v1/review/reply:
        post:
            tags:
                - Review
            description: B端回复评价
            operationId: Review_ReplyReview
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.review.v1.ReplyReviewRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.ReplyReviewReply'
    /v1/review/replyupdate:
        put:
            tags:
                - Review
            description: B端修改回复
            operationId: Review_ReplyReviewUpdate
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.review.v1.ReplyReviewUpdateRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.ReplyReviewUpdateReply'
...
    /v1/update/{reviewID}:
        put:
            tags:
                - Review
            description: c端更新评价
            operationId: Review_UpdateReview
            parameters:
                - name: reviewID
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.review.v1.UpdateReviewRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.UpdateReviewReply'
    /v1/{storeID}/reviews:
        get:
            tags:
                - Review
            description: 根据商家ID查询评价列表
            operationId: Review_ListReviewByStoreID
            parameters:
                - name: storeID
                  in: path
                  required: true
                  schema:
                    type: string
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: size
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.ListReviewByStoreIDReply'
    v1/appeal/audit:
        post:
            tags:
                - Review
            description: O端审核申诉
            operationId: Review_AuditAppeal
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.review.v1.AuditAppealRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.AuditAppealReply'
    v1/review/appeal:
        post:
            tags:
                - Review
            description: B端申诉评价
            operationId: Review_AppealReview
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.review.v1.AppealReviewRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.AppealReviewReply'
    v1/review/audit:
        post:
            tags:
                - Review
            description: O端审核评价
            operationId: Review_AuditReview
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.review.v1.AuditReviewRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.AuditReviewReply'
    v1/review/{reviewID}:
        get:
            tags:
                - Review
            description: c端获取评价详情
            operationId: Review_GetReview
            parameters:
                - name: reviewID
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.GetReviewReply'
    v1/reviews/{userID}:
        get:
            tags:
                - Review
            description: c端获取自己的全部评价详情
            operationId: Review_GetAllReviews
            parameters:
                - name: userID
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.GetAllReviewsReply'
components:
    schemas:
        api.review.v1.AllReviewsInfo:
            type: object
            properties:
                userID:
                    type: string
                orderID:
                    type: string
                reviewID:
                    type: string
                score:
                    type: integer
                    format: int32
                serviceScore:
                    type: integer
                    format: int32
                expressScore:
                    type: integer
                    format: int32
                content:
                    type: string
                picInfo:
                    type: string
                videoInfo:
                    type: string
                status:
                    type: integer
                    format: int32
        api.review.v1.AppealReviewReply:
            type: object
            properties:
                appealID:
                    type: string
        api.review.v1.AppealReviewRequest:
            type: object
            properties:
                reviewID:
                    type: string
                storeID:
                    type: string
                reason:
                    type: string
                content:
                    type: string
                picInfo:
                    type: string
                videoInfo:
                    type: string
            description: B端申诉评价
        api.review.v1.AuditAppealReply:
            type: object
            properties: {}
        api.review.v1.AuditAppealRequest:
            type: object
            properties:
                appealID:
                    type: string
                status:
                    type: integer
                    format: int32
                opUser:
                    type: string
                opReason:
                    type: string
                opRemarks:
                    type: string
            description: O端审核申诉
        api.review.v1.AuditReviewReply:
            type: object
            properties: {}
        api.review.v1.AuditReviewRequest:
            type: object
            properties:
                reviewID:
                    type: string
                status:
                    type: integer
                    format: int32
                opUser:
                    type: string
                opReason:
                    type: string
                opRemarks:
                    type: string
            description: O端审核评价
        api.review.v1.CreateReviewReply:
            type: object
            properties:
                reviewID:
                    type: string
        api.review.v1.CreateReviewRequest:
            type: object
            properties:
                userID:
                    type: string
                orderID:
                    type: string
                storeID:
                    type: string
                score:
                    type: integer
                    format: int32
                serviceScore:
                    type: integer
                    format: int32
                expressScore:
                    type: integer
                    format: int32
                content:
                    type: string
                picInfo:
                    type: string
                videoInfo:
                    type: string
                anonymous:
                    type: boolean
            description: c端创建评价
        api.review.v1.DeleteReviewReply:
            type: object
            properties:
                rowaffected:
                    type: string
        api.review.v1.GetAllReviewsReply:
            type: object
            properties:
                data:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.review.v1.AllReviewsInfo'
        api.review.v1.GetReviewReply:
            type: object
            properties:
                userID:
                    type: string
                orderID:
                    type: string
                score:
                    type: integer
                    format: int32
                serviceScore:
                    type: integer
                    format: int32
                expressScore:
                    type: integer
                    format: int32
                content:
                    type: string
                picInfo:
                    type: string
                videoInfo:
                    type: string
                status:
                    type: integer
                    format: int32
...
        api.review.v1.ListReviewByStoreIDReply:
            type: object
            properties:
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.review.v1.ReviewInfo'
        api.review.v1.ReplyReviewReply:
            type: object
            properties:
                replyID:
                    type: string
        api.review.v1.ReplyReviewRequest:
            type: object
            properties:
                reviewID:
                    type: string
                storeID:
                    type: string
                content:
                    type: string
                picInfo:
                    type: string
                videoInfo:
                    type: string
            description: B端回复评价
        api.review.v1.ReplyReviewUpdateReply:
            type: object
            properties:
                replyID:
                    type: string
                rowsaffected:
                    type: string
        api.review.v1.ReplyReviewUpdateRequest:
            type: object
            properties:
                reviewID:
                    type: string
                storeID:
                    type: string
                replyID:
                    type: string
                content:
                    type: string
                picInfo:
                    type: string
                videoInfo:
                    type: string
            description: B端修改回复
        api.review.v1.ReviewInfo:
            type: object
            properties:
                userID:
                    type: string
                orderID:
                    type: string
                score:
                    type: integer
                    format: int32
                serviceScore:
                    type: integer
                    format: int32
                expressScore:
                    type: integer
                    format: int32
                content:
                    type: string
                picInfo:
                    type: string
                videoInfo:
                    type: string
                status:
                    type: integer
                    format: int32
        api.review.v1.UpdateReviewReply:
            type: object
            properties:
                reviewID:
                    type: string
                rowaffected:
                    type: string
        api.review.v1.UpdateReviewRequest:
            type: object
            properties:
                reviewID:
                    type: string
                userID:
                    type: string
                score:
                    type: integer
                    format: int32
                serviceScore:
                    type: integer
                    format: int32
                expressScore:
                    type: integer
                    format: int32
                content:
                    type: string
                picInfo:
                    type: string
                videoInfo:
                    type: string
            description: c端更新评价
...